requirements.txt
asgiref==3.7.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
Django==4.2.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.3
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathy==0.10.1
praw==7.7.0
prawcore==2.3.0
preshed==3.0.8
pydantic==1.10.9
python-dateutil==2.8.2
pytrends==4.9.2
pytz==2023.3
rake-nltk==1.0.6
regex==2023.6.3
requests==2.31.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
thinc==8.1.10
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
update-checker==0.18.0
urllib3==2.0.3
wasabi==1.1.2
websocket-client==1.5.3
google-ads==21.2.0
google-auth
google-auth-oauthlib
google-auth-httplib2